[refactor] 계산기 패키지 이동
