+++ src/main/java/study/calculator/StringCalculator.java
@@ -1,4 +1,4 @@
-package study;
+package study.calculator;
 
 /**
  * Created by wyparks2@gmail.com on 2019-06-15
