+++ src/test/java/study/calculator/CalculatorOperatorTest.java
@@ -1,4 +1,4 @@
-package study;
+package study.calculator;
 
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
