+++ src/test/java/study/calculator/StringCalculatorTest.java
@@ -1,4 +1,4 @@
-package study;
+package study.calculator;
 
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
